server/requirements.txt
mariadb==1.1.10
packaging==24.1
paho-mqtt==2.1.0
